skills/observability-harness/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "observability-harness-skill"
version = "1.0.0"
description = "Standalone observability harness runtime for Victoria baseline and smoke validation"
requires-python = ">=3.10, <3.14"
dependencies = [
    "jsonschema>=4.23.0",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]

[tool.uv]
package = false
